TrackerSelectionScreen -> MVVM
